+++ main/src/main/java/net/oneandone/stool/cli/Main.java
@@ -91,7 +91,6 @@ public class Main {
                     cli.add(Select.class, "select -fuzzy=@select.fuzzy name?=null");
                     cli.base(ProjectCommand.class, "-autorestart=@auto.restart -autostop=@auto.stop -stage=null -all -fail "
                             + "{ setAutoRestart(autorestart) setAutoStop(autostop) setStage(stage) setAll(all) setFail(fail) }");
-                      cli.add(Build.class, "build -here command*");
                       cli.add(Cd.class, "cd target?=null { setTarget(target) }");
                       cli.add(Cleanup.class, "cleanup");
                       cli.add(Config.class, "config property* { property*(property) }");
@@ -99,7 +98,7 @@ public class Main {
                       cli.add(Ls.class, "list -defaults=@list.defaults info* { select*(info) }");
                       cli.add(Move.class, "move dest");
                       cli.add(Port.class, "port port { port*(port) }");
-                      cli.add(Refresh.class, "refresh -build=@refresh.build -restore");
+                      cli.add(Refresh.class, "refresh -restore");
                       cli.add(Remove.class, "remove -batch -force -backstage { setBackstage(backstage) }");
                       cli.add(Restart.class, "restart -nocache");
                       cli.add(Start.class, "start -tail -nocache");