+++ main/src/main/java/net/oneandone/stool/configuration/StageConfiguration.java
@@ -50,9 +50,6 @@ public class StageConfiguration {
     @Option(key = "prepare")
     public String prepare;
 
-    @Option(key = "build")
-    public String build;
-
     @Option(key = "refresh")
     public String refresh;
 
@@ -101,7 +98,6 @@ public class StageConfiguration {
     public StageConfiguration(String mavenHome, FileNode template, String refresh) {
         this.name = "noname";
         this.prepare = "";
-        this.build = "false";
         this.refresh = refresh;
         this.notify = new ArrayList<>();
         this.notify.add(NOTIFY_CREATED_BY);