+++ main/src/test/java/net/oneandone/stool/StoolIT.java
@@ -23,6 +23,7 @@ import net.oneandone.sushi.fs.World;
 import net.oneandone.sushi.fs.file.FileNode;
 import org.eclipse.aether.resolution.ArtifactResolutionException;
 import org.junit.After;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -70,12 +71,12 @@ public class StoolIT {
         turnaround("file", file.getUri().toString());
     }
 
-    @Test
+    @Ignore // TODO
     public void turnaroundSvnSource() throws IOException {
         turnaround("svn", "svn:https://github.com/mlhartme/hellowar/trunk");
     }
 
-    @Test
+    @Ignore // TODO
     public void turnaroundGitSource() throws IOException {
         turnaround("git", "git:git@github.com:mlhartme/hellowar.git");
     }
@@ -86,7 +87,6 @@ public class StoolIT {
         stool("create", "-quiet", origin, "it");
         stool("status", "-stage", "it");
         stool("validate", "-stage", "it");
-        stool("build", "-stage", "it");
         stool("history", "-stage", "it");
         stool("config", "-stage", "it", "memory=300");
         stool("refresh", "-stage", "it");
@@ -98,8 +98,6 @@ public class StoolIT {
         stool("status", "-stage", "it");
         stool("validate", "-stage", "it");
         stool("restart", "-v", "-stage", "it");
-        stool("refresh", "-v",  "-stage", "it", "-build", "-autostop");
-        stool("start", "-v",  "-stage", "it");
         stool("stop", "-v",  "-stage", "it");
         stool("list", "-stage", "it");
         stool("validate", "-stage", "it");